fix: overlayroot を使っていない場合にエラーが出ないようにした

+++ script/sync_overlay.zsh
@@ -1,5 +1,7 @@
 #!/usr/bin/zsh
 
-sudo mount -o remount,rw /media/root-ro
-sudo cp -f /var/spool/cron/crontabs/root /media/root-ro/var/spool/cron/crontabs/root
-sudo mount -o remount,ro /media/root-ro
+if [ -e /media/root-ro ]; then
+    sudo mount -o remount,rw /media/root-ro
+    sudo cp -f /var/spool/cron/crontabs/root /media/root-ro/var/spool/cron/crontabs/root
+    sudo mount -o remount,ro /media/root-ro
+fi
